+++ src/storage.ts
@@ -406,6 +406,20 @@ export class Storage {
    */
   public getSessionTokenPath(label: string): string {
     return label
+      ? path.join(this.globalStorageUri.fsPath, label, "session")
+      : path.join(this.globalStorageUri.fsPath, "session")
+  }
+
+  /**
+   * Return the directory for the deployment with the provided label to where
+   * its session token was stored by older code.
+   *
+   * If the label is empty, read the old deployment-unaware config instead.
+   *
+   * The caller must ensure this directory exists before use.
+   */
+  public getLegacySessionTokenPath(label: string): string {
+    return label
       ? path.join(this.globalStorageUri.fsPath, label, "session_token")
       : path.join(this.globalStorageUri.fsPath, "session_token")
   }
@@ -435,10 +449,10 @@ export class Storage {
   /**
    * Configure the CLI for the deployment with the provided label.
    *
-   * Falsey values are a no-op; we avoid unconfiguring the CLI to avoid breaking
-   * existing connections.
+   * Falsey URLs and null tokens are a no-op; we avoid unconfiguring the CLI to
+   * avoid breaking existing connections.
    */
-  public async configureCli(label: string, url: string | undefined, token: string | undefined | null) {
+  public async configureCli(label: string, url: string | undefined, token: string | null) {
     await Promise.all([this.updateUrlForCli(label, url), this.updateTokenForCli(label, token)])
   }
 
@@ -459,15 +473,15 @@ export class Storage {
   /**
    * Update the session token for a deployment with the provided label on disk
    * which can be used by the CLI via --session-token-file.  If the token is
-   * falsey, do nothing.
+   * null, do nothing.
    *
    * If the label is empty, read the old deployment-unaware config instead.
    */
   private async updateTokenForCli(label: string, token: string | undefined | null) {
-    if (token) {
+    if (token !== null) {
       const tokenPath = this.getSessionTokenPath(label)
       await fs.mkdir(path.dirname(tokenPath), { recursive: true })
-      await fs.writeFile(tokenPath, token)
+      await fs.writeFile(tokenPath, token ?? "")
     }
   }
 
@@ -485,6 +499,22 @@ export class Storage {
     return {
       url: url.status === "fulfilled" ? url.value.trim() : "",
       token: token.status === "fulfilled" ? token.value.trim() : "",
+    }
+  }
+
+  /**
+   * Migrate the session token file from "session_token" to "session", if needed.
+   */
+  public async migrateSessionToken(label: string) {
+    const oldTokenPath = this.getLegacySessionTokenPath(label)
+    const newTokenPath = this.getSessionTokenPath(label)
+    try {
+      await fs.rename(oldTokenPath, newTokenPath)
+    } catch (error) {
+      if ((error as NodeJS.ErrnoException)?.code === "ENOENT") {
+        return
+      }
+      throw error
     }
   }